[COLAB-2439] Move user tracking services to parent folder
